Refactor reused enum-based macros
